(SIMP-3636) Convert compliance profiles to JSON (#23)
* Prior to this commit all values in the compliance profiles were strings.
This commit uses syntax to define each value's type.

* JSON is generally faster to deserialize than YAML, and it is more
strict.  Given the size of the compliance profiles, and the associated
potential for errors, this commit converts them from YAML to JSON,
updating the `hiera.yaml` to use the `json` backend so the change is
transparent to the user.  Additionally, this commit includes two scripts
in the `utils/` directory to convert the files back and forth between
the two formats.

SIMP-3636 #close
SIMP-3627 #close