(SIMP-3784) Replace the compliance mapper's report generator (#26)
Replace the compliance mapper's report generator parsing code
with the shared library used by the enforcement backend. This
means that the code to read and parse v1 format compliance maps
is now centralized.

Add vendored (maps-in-modules) support to the compiler for v1
format maps. Note: the operation of this is not well defined, and
intended for use in the v2 format of the compliance maps.

Also expand the shared library to a real object tree, with defined
api endpoints, like load() to load a set of compliance maps, and
list_puppet_params() to generate a puppet parameter hash.

Note: this code was tested using the existing spec tests without
modifications to ensure that the change is a drop-in replacement.
The code coverage of the tests may not be complete to do this.

SIMP-3784 #close